+++ src/main/java/dev/orne/config/DefaultConfigProvider.java
@@ -1,0 +1,127 @@
+/**
+ * 
+ */
+package dev.orne.config;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import org.apache.commons.lang3.Validate;
+
+/**
+ * Default implementation of {@code ConfigProvider}.
+ * 
+ * @version 1.0
+ * @author (w) Iker Hernaez<i.hernaez@hif-soft.net>
+ * @since 1.0, 2019-07
+ */
+public class DefaultConfigProvider
+implements ConfigProvider {
+
+	/** Error message for null default configuration. */
+	private static final String NULL_DEFAULT_ERR =
+			"Default config is required.";
+	/** Error message for null configuration register attempt. */
+	private static final String NULL_CONFIG_ERR =
+			"Cannot register a null configuration.";
+	/** Error message for null preferred configuration. */
+	private static final String NULL_PREFERRED_ERR =
+			"Unexpected null preferred configuration for class ";
+
+	/** The default configuration. */
+	private final Config defaultConfig;
+	/** The available configurations mappings. */
+	private final Map<Class<?>, Config> mappings;
+
+	/**
+	 * Creates a new instance.
+	 * 
+	 * @param defaultConfig The default configuration
+	 */
+	public DefaultConfigProvider(
+			final Config defaultConfig) {
+		super();
+		Validate.notNull(defaultConfig, NULL_DEFAULT_ERR);
+		this.defaultConfig = defaultConfig;
+		this.mappings = new HashMap<Class<?>, Config>();
+		mapConfigType(defaultConfig.getClass(), defaultConfig);
+	}
+
+	/**
+	 * Registers a new configuration available for bean configuration.
+	 * 
+	 * @param config The configuration to register
+	 */
+	public void registerConfig(final Config config) {
+		Validate.notNull(config, NULL_CONFIG_ERR);
+		mapConfigType(config.getClass(), config);
+	}
+
+	/**
+	 * Maps configuration type and all its interfaces to the instance
+	 * passed as argument. If the type has super class different of
+	 * {@code Object} repeats the mapping for it.
+	 * 
+	 * @param type The type of configuration to scan
+	 * @param config The configuration instance
+	 */
+	protected void mapConfigType(
+			final Class<?> type,
+			final Config config) {
+		if (!this.mappings.containsKey(type)) {
+			this.mappings.put(type, config);
+			for (final Class<?> iface : type.getInterfaces()) {
+				mapConfigType(iface, config);
+			}
+			final Class<?> superType = type.getSuperclass();
+			if (superType != null && !Object.class.equals(superType)) {
+				mapConfigType(superType, config);
+			}
+		}
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public Config getDefaultConfig() {
+		return this.defaultConfig;
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public Config selectConfig(
+			final ConfigurationOptions options,
+			final Class<?> targetClass) {
+		Config result = null;
+		if (options == null || options.preferedConfigs() == null) {
+			result = this.defaultConfig;
+		} else {
+			for (final Class<?> prefered : options.preferedConfigs()) {
+				Validate.notNull(prefered, NULL_PREFERRED_ERR + targetClass);
+				if (this.mappings.containsKey(prefered)) {
+					result = this.mappings.get(prefered);
+					break;
+				}
+			}
+			if (result == null && options.fallbackToDefaultConfig()) {
+				result = this.defaultConfig;
+			}
+		}
+		return result;
+	}
+
+	/**
+	 * Returns {@code true} if a configuration is mapped for the requested
+	 * class.
+	 * 
+	 * @param type The requested class
+	 * @return {@code true} if a configuration is mapped for the requested
+	 * class
+	 */
+	protected boolean isMapped(final Class<?> type) {
+		return this.mappings.containsKey(type);
+	}
+}